изменена структура проекта
