+++ 05.homework/CMakeLists.txt
@@ -8,6 +8,5 @@ set_target_properties(random_shuffle PROPERTIES CXX_STANDARD 17 CXX_STANDARD_REQ
 add_executable(chrono_example chrono_example.cpp)
 set_target_properties(chrono_example PROPERTIES CXX_STANDARD 17 CXX_STANDARD_REQUIRED ON)
 
-add_executable(statistics statistics.cpp)
-set_target_properties(statistics PROPERTIES CXX_STANDARD 17 CXX_STANDARD_REQUIRED ON)
+add_subdirectory(statistics)
